+++ strax/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-__version__ = "2.0.2"
+__version__ = "2.1.3"
 
 # Glue the package together
 # See https://www.youtube.com/watch?v=0oTh1CXRaQ0 if this confuses you
@@ -25,7 +25,6 @@ from .processor import *
 from .processors import *
 from .context import *
 from .run_selection import *
-from .corrections import *
 
 from .io import *