+++ strax/dtypes.py
@@ -155,6 +155,8 @@ def hitlet_with_data_dtype(n_samples=2):
         (("Dummy max_gap required for splitting", "max_gap"), np.int32),
         (("Dummy max_diff required for splitting", "max_diff"), np.int32),
         (("Dummy min_diff required for splitting", "min_diff"), np.int32),
+        (("Dummy first_channel required for splitting", "first_channel"), np.int16),
+        (("Dummy last_channel required for splitting", "last_channel"), np.int16),
         (("Maximum interior goodness of split", "max_goodness_of_split"), np.float32),
     ]
 
@@ -179,9 +181,12 @@ def peak_dtype(
         # For peaklets this is likely to be overwritten:
         (("Classification of the peak(let)", "type"), np.int8),
         (("Integral across channels [PE]", "area"), np.float32),
+        (("Fraction of area seen by the top array", "area_fraction_top"), np.float32),
         (("Integral per channel [PE]", "area_per_channel"), np.float32, n_channels),
         (("Number of hits contributing at least one sample to the peak ", "n_hits"), np.int32),
         (("Waveform data in PE/sample (not PE/ns!)", "data"), np.float32, n_sum_wv_samples),
+        (("Weighted average center time of the peak [ns]", "center_time"), np.int64),
+        (("Weighted relative median time of the peak [ns]", "median_time"), np.float32),
         (("Peak widths in range of central area fraction [ns]", "width"), np.float32, n_widths),
         (
             ("Peak widths: time between nth and 5th area decile [ns]", "area_decile_from_midpoint"),
@@ -203,6 +208,17 @@ def peak_dtype(
             (
                 ("Smallest time difference between apexes of hits inside peak [ns]", "min_diff"),
                 np.int32,
+            ),
+            (
+                (
+                    "First channel/PMT number inside peak (sorted by apexes of hits)",
+                    "first_channel",
+                ),
+                np.int16,
+            ),
+            (
+                ("Last channel/PMT number inside peak (sorted by apexes of hits)", "last_channel"),
+                np.int16,
             ),
         ]
     if store_data_top: